202510122_박광열 : 출하 단위 변경으로 인한 리팩터링 (int -> double)

+++ src/main/java/mes/app/shipment/ShipmentDoaController.java
@@ -115,7 +115,7 @@ public class ShipmentDoaController {
 				}
 				ShipmentStatus status = ShipmentStatus.SHIPPED;
 				smh.setState(status.getLabel());
-				smh.setTotalQty((float)orderSum);
+				smh.setTotalQty((double)orderSum);
 				smh.set_audit(user);
 				smh = this.shipmentHeadRepository.save(smh);
 			}
@@ -171,7 +171,7 @@ public class ShipmentDoaController {
 				}
 			}
 			smh.setState("shipped");
-			smh.setTotalQty((float)orderSum);
+			smh.setTotalQty((double)orderSum);
 			smh.set_audit(user);
 			smh = this.shipmentHeadRepository.save(smh);
 		}
